% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rHP2.R
\name{rHP}
\alias{rHP}
\title{Hawkes Process simulation with different algorithms}
\usage{
rHP(
  n,
  T,
  lam0,
  phi,
  Lam0,
  Phi,
  Lam0Inv,
  PhiInv,
  method = c("Thinning", "Numerical Inversion", "cf Inversion")
)
}
\arguments{
\item{n}{The number of independent replications for the Hawkes Process simulation}

\item{T}{The simulation duration}

\item{lam0}{The base intensity function of Hawkes Process}

\item{phi}{The excitation function of Hawkes Process}

\item{Lam0}{The cumulative base intensity function of Hawkes}

\item{Phi}{The cumulative excitation function of Hawkes Process}

\item{Lam0Inv}{The inverse cumulative base intensity function of Hawkes Process}

\item{PhiInv}{The inverse cumulative excitation function of Hawkes Process}

\item{method}{The simulation method, "Thinning", "Numerical Inversion","cf Inversion"}
}
\value{
The event times of n replications and the replication number of each event
}
\description{
This function simulates Hawkes Process with

1- Ogata's modified thinning algorithm (Ogata, 1981, p.25, Algorithm 2)

2- Generating Hawkes Process by clusters (Hawkes and Oakes,1974)

The clusters of Hawkes Process, the Non-Homogeneous Poisson Processes are simulated with

an proposed algorithm that combines, inversion, order statistics and transformation method.

Two inversion methods are proposed

a- The closed form of the inverse cumulative intensity function is available ("cf Inversion")

b- The closed form of the inverse cumulative intensity function is not available ("Numerical Inversion")

In the later case, the numerical inversion is used
}
\examples{
rHP(n=10000,T=1000,lam0=function(x) {2000*(x+250)^(-2.33)},phi=function(x) {2000*(x+250)^(-2.33)},Lam0=NULL,Phi=NULL,Lam0Inv=NULL,PhiInv=NULL,method="Thinning")
rHP(n=10000,T=1000,lam0=function(x) {2000*(x+250)^(-2.33)},phi=function(x) {2000*(x+250)^(-2.33)},Lam0=NULL,Phi=NULL,Lam0Inv=NULL,PhiInv=NULL,method="Numerical Inversion")
rHP(n=10000,T=1000,lam0=function(x) {2000*(x+250)^(-2.33)},phi=function(x) {2000*(x+250)^(-2.33)},Lam0=function(x) {(-2000/1.33)*(x+250)^(-1.33)+(2000/1.33*250^(-1.33))},Phi=function(x) {(-2000/1.33)*(x+250)^(-1.33)+(2000/1.33*250^(-1.33))},Lam0Inv=function(x) {(-(x-(2000/1.33*250^(-1.33)))*(1.33/2000))^(-(1/1.33))-250},PhiInv=function(x) {(-(x-(2000/1.33*250^(-1.33)))*(1.33/2000))^(-(1/1.33))-250},method="cf Inversion")

}
