% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rNHPP2.R
\name{rNHPP}
\alias{rNHPP}
\title{Non-Homogeneous Poisson Process simulation with different algorithms}
\usage{
rNHPP(
  n,
  T,
  lam0,
  Lam0,
  Lam0Inv,
  method = c("Thinning", "cf Inversion", "Numerical Inversion")
)
}
\arguments{
\item{n}{The number of independent replications for the Hawkes Process simulation}

\item{T}{The simulation duration}

\item{lam0}{The base intensity function of Hawkes Process}

\item{Lam0}{The cumulative base intensity function of Hawkes}

\item{Lam0Inv}{The inverse cumulative base intensity function of Hawkes Process}

\item{method}{The simulation method, "Thinning", "Numerical Inversion","cf Inversion"}
}
\value{
The event times of n replications and the replication number of each event
}
\description{
This function simulates Non-Homogeneous Poisson Process with

1- Ogata's modified thinning algorithm (Ogata, 1981, p.24, Algorithm 1)

2- A proposed algorithm that combines Inversion, Order Statistics and Transformation method

Two inversion methods are proposed

a- The closed form of the inverse cumulative intensity function is available ("cf Inversion")

b- The closed form of the inverse cumulative intensity function is not available ("Numerical Inversion")

In the later case, the numerical inversion is used
}
\examples{
rNHPP(n=10000,T=1000,lam0=function(x) {2000*(x+250)^(-2.33)},Lam0=NULL,Lam0Inv=NULL,method="Thinning")
rNHPP(n=10000,T=1000,lam0=function(x) {2000*(x+250)^(-2.33)},Lam0=NULL,Lam0Inv=NULL,method="Numerical Inversion")
rNHPP(n=10000,T=1000,lam0=function(x) {2000*(x+250)^(-2.33)},Lam0=function(x) {(-2000/1.33)*(x+250)^(-1.33)+(2000/1.33*250^(-1.33))},Lam0Inv=function(x) {(-(x-(2000/1.33*250^(-1.33)))*(1.33/2000))^(-(1/1.33))-250},method="cf Inversion")

}
